Create bucket attempt before qldb create
This switches the order of the environment init
to attempt to create the S3 bucket before the ledger.
This is done since the S3 bucket creation is far more
likely to fail due to global name conflicts, and the
ledger creation will almost certainly succeed, or be
unnecessary if you have access.

Signed-off-by: Kevin O'Donnell <kevin@blockchaintp.com>